Replace OperatorsSpec with OperatorsTests.
